Create s3-file-info.py
This is a simple python3 script that will list the contents of the root of an S3 bucket.  

I'm not a developer by training.  So, I'm not really good with PEP8 or even DRY standards.  

I've got a project that I'm working on where I want to process files in S3 but realized that I have no idea how to do it.  Ironically, searching the Internet was less than helpful.  Since I'm not a developer, I'll put code snippets here that work for me.  

Please, if you can make this prettier, more readable, or improve on it in general, let me know.  I love learning new things and making things better.